requirements.txt
djangorestframework==3.9.4
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.1
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
Pillow==6.0.0
psycopg2==2.8.2
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.2
